+++ lib/src/handler/w3c/session.dart
@@ -1,3 +1,5 @@
+import 'package:appium_driver/src/common/request.dart';
+import 'package:appium_driver/src/common/w3c/command.dart';
 import 'package:appium_driver/src/common/webdriver_handler.dart';
 
 import 'package:webdriver/src/common/request.dart'; // ignore: implementation_imports
@@ -30,4 +32,15 @@ class W3cSessionHandler extends SessionHandler {
   @override
   SessionInfo parseInfoResponse(WebDriverResponse response) =>
       SessionInfo(response.body, WebDriverSpec.W3c, Capabilities.empty);
+
+  @override
+  WebDriverRequest buildGetCapabilitiesRequest() {
+    return AppiumWebDriverRequest.sendRequest(W3CCommands.GET_CAPABILITIES);
+  }
+
+  @override
+  Map<String, dynamic> parseGetCapabilitiesResponse(
+      WebDriverResponse response) {
+    return parseW3cResponse(response);
+  }
 }