+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 - Add get sessions command: `driver.sessions.get()`
 - Add get status: `driver.status.get()`
+- Add get capabilities: `driver.session.getCapabilities()`
 
 ## [0.0.6] - 2019-06-30
 - Add device lock related commands: `driver.device.*`